MFC r304920:
In ncurses baudrate definitions, avoid warnings about implicit
conversions from int to short changing the values.  This applies to
B38400 and higher, since their values do not fit into a short.

However, since the wrapped values are still unique, and they only serve
as keys, there is no problem in adding a cast to silence the warnings.
This also avoids changing the ABI, which would happen if we changed
NCURSES_OSPEED to int.

Discussed with:	Thomas Dickey


git-svn-id: https://svn.freebsd.org/base/stable/11@305547 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f