Give the full syntax of the 'count' arg for all commmands that support it.  This arg is most interesting for the 'break' command where it never worked, and for the step command where it is powerful but too fragile to use much.
Give the full syntax of the 'addr' arg for these commands and some
others.  Rename it from 'address' for the generic command.

Fix description of how 'count' is supposed to work for the 'break'
command.

Don't (mis)describe the syntax of the comma for the 'step' command.

Expand the description for the generic command.

Give the full syntax for the 'examine' command.  It was also missing
the possible values for the modifier.

Fix mdoc syntax error for the 'search' command.

Remove FUD about consequences of not having a trap handler for the
'search' command.


git-svn-id: https://svn.freebsd.org/base/head@305661 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f