Disable SYSTEM_COMPILER by default.
This is a direct commit to releng/11.0.

Having it enabled can lead to a situation where building
on one system and installing on another will fail due
to not finding cc in the OBJDIR.

An actual fix will be made on head separately.

PR:		212877
Relnotes:	yes
Sponsored by:	Dell EMC Isilon
Approved by:	re (gjb)


git-svn-id: https://svn.freebsd.org/base/releng/11.0@306143 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f