Avoid depending on the <sys/endian.h> header for le*enc and be*enc. Not only is the header unportable, the encoding/decoding functions are as well.  Instead, duplicate the handful of small inlines we need into a private header called endian.h.
Aside: an alternative approach is to move the encoding/decoding
functions to a separate system header.  While the header is still
nonportable, such an approach would make it possible to re-use the
definitions by playing games with include paths. This may be the
preferred approach if more (build) utilities need this.  This
change does not preclude that.  In fact, it makes it easier.


git-svn-id: https://svn.freebsd.org/base/head@306330 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f